requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24.2
psycopg[binary]>=3.2
celery[redis]>=5.4
redis>=5.0
python-json-logger>=2.0
dj-database-url>=2.2
sentry-sdk>=2.14
gunicorn>=22.0

pytest>=8.3
pytest-django>=4.9
factory-boy>=3.3
faker>=30.0
freezegun>=1.5
ruff>=0.6
mypy>=1.11
